requirements.txt
asgiref==3.2.7
django-rest-auth==0.9.5
Django==3.0.5
djangorestframework==3.11.0
factory-boy<=2.12.0
Faker<=3.0.0
pre-commit
psycopg2==2.8.5
pylint
pytest==5.4.1
pytz==2019.3
djangorestframework-simplejwt==4.4.0
sqlparse==0.3.1
